+++ src/fill/into.rs
@@ -1,8 +1,8 @@
-use super::Expand;
+use super::Fill;
 use a1_notation::A1;
 
 #[allow(clippy::from_over_into)]
-impl Into<A1> for Expand {
+impl Into<A1> for Fill {
     fn into(self) -> A1 {
         a1_notation::row_range(self.start_row, self.end_row().shift_up(1))
     }
@@ -14,7 +14,7 @@ mod tests {
 
     #[test]
     fn into_a1_finite() {
-        let a1: A1 = Expand {
+        let a1: A1 = Fill {
             amount: Some(3),
             start_row: 0.into(),
         }
@@ -24,7 +24,7 @@ mod tests {
 
     #[test]
     fn into_a1_infinite() {
-        let a1: A1 = Expand {
+        let a1: A1 = Fill {
             amount: None,
             start_row: 0.into(),
         }